Updated measurements - Moved all measurement functions to measurements.py - Updated README.md
